refactor Reports CSV/PDF code
