+++ app/services/report/csv/member.rb
@@ -1,21 +1,8 @@
 module Report
   module Csv
-    class Member
+    class Member < Base
       def initialize(org, collection)
-        @collection = collection
-        @decorator = MemberReportDecorator.new(org, @collection)
-      end
-
-      def name
-        @decorator.name(:csv)
-      end
-
-      def mime_type
-        Report::Csv::MIME_TYPE
-      end
-
-      def run
-        Report::Csv.run(@decorator.headers, @decorator.rows)
+        @decorator = MemberReportDecorator.new(org, collection)
       end
     end
   end
