+++ app/services/report/pdf/member.rb
@@ -1,21 +1,8 @@
 module Report
   module Pdf
-    class Member
+    class Member < Base
       def initialize(org, collection)
-        @collection = collection
-        @decorator = MemberReportDecorator.new(org, @collection)
-      end
-
-      def name
-        @decorator.name(:pdf)
-      end
-
-      def mime_type
-        Report::Pdf::MIME_TYPE
-      end
-
-      def run
-        Report::Pdf.run(@decorator.headers, @decorator.rows)
+        @decorator = MemberReportDecorator.new(org, collection)
       end
     end
   end
